+++ bluepysnap/edges/edges.py
@@ -21,7 +21,6 @@ import numpy as np
 
 from bluepysnap._doctools import AbstractDocSubstitutionMeta
 from bluepysnap.circuit_ids import CircuitEdgeIds, CircuitNodeIds
-from bluepysnap.circuit_ids_types import CircuitNodeId
 from bluepysnap.edges.edge_population import EdgePopulation
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.network import NetworkObject
@@ -101,7 +100,7 @@ class Edges(
         if edge_ids is None:
             raise BluepySnapError("You need to set edge_ids in get.")
         if properties is None:
-            return edge_ids
+            return self.ids(edge_ids)
         return super().get(edge_ids, properties)
 
     def afferent_nodes(self, target, unique=True):
@@ -221,39 +220,6 @@ class Edges(
             source=source_node_id, target=target_node_id, properties=properties
         )
 
-    @staticmethod
-    def _add_circuit_ids(its, source, target):
-        """Generator comprehension adding the CircuitIds to the iterator.
-
-        Notes:
-            Using closures or lambda functions would result in override functions and so the
-            source and target would be the same for all the populations.
-        """
-        return (
-            (CircuitNodeId(source, source_id), CircuitNodeId(target, target_id), count)
-            for source_id, target_id, count in its
-        )
-
-    @staticmethod
-    def _add_edge_ids(its, source, target, pop_name):
-        """Generator comprehension adding the CircuitIds to the iterator."""
-        return (
-            (
-                CircuitNodeId(source, source_id),
-                CircuitNodeId(target, target_id),
-                CircuitEdgeIds.from_dict({pop_name: edge_id}),
-            )
-            for source_id, target_id, edge_id in its
-        )
-
-    @staticmethod
-    def _omit_edge_count(its, source, target):
-        """Generator comprehension adding the CircuitIds to the iterator."""
-        return (
-            (CircuitNodeId(source, source_id), CircuitNodeId(target, target_id))
-            for source_id, target_id in its
-        )
-
     def iter_connections(
         self, source=None, target=None, return_edge_ids=False, return_edge_count=False
     ):
@@ -276,21 +242,13 @@ class Edges(
             raise BluepySnapError(
                 "`return_edge_count` and `return_edge_ids` are mutually exclusive"
             )
-        for name, pop in self.items():
-            it = pop.iter_connections(
+        for pop in self.values():
+            yield from pop.iter_connections(
                 source=source,
                 target=target,
                 return_edge_ids=return_edge_ids,
                 return_edge_count=return_edge_count,
             )
-            source_pop = pop.source.name
-            target_pop = pop.target.name
-            if return_edge_count:
-                yield from self._add_circuit_ids(it, source_pop, target_pop)
-            elif return_edge_ids:
-                yield from self._add_edge_ids(it, source_pop, target_pop, name)
-            else:
-                yield from self._omit_edge_count(it, source_pop, target_pop)
 
     def __getstate__(self):
         """Make Edges pickle-able, without storing state of caches."""