+++ bluepysnap/query.py
@@ -1,4 +1,5 @@
 """Module to process search queries of nodes/edges."""
+
 import operator
 from collections.abc import Mapping
 from copy import deepcopy
@@ -125,7 +126,7 @@ def _properties_mask(data, population_name, queries):
             prop_mask = np.logical_and(prop >= v1, prop <= v2)
         elif isinstance(values, Mapping):
             prop_mask = _complex_query(prop, values)
-        elif isinstance(values, list):
+        elif utils.is_iterable(values):
             prop_mask = prop.isin(values)
         else:
             prop_mask = prop == values